requirements.txt
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
pandas>=1.5.0
folium>=0.14.0
matplotlib>=3.6.0
requests>=2.28.0
fpdf2>=2.7.9
